Added rocks and collision.

+++ src/main/java/com/heychinaski/ld23/Input.java
@@ -58,8 +58,8 @@ public class Input {
     this.mouseX = mouseX;
     this.mouseY = mouseY;
     
-    this.worldMouseX = mouseX + (camera.x - (camera.width / 2));
-    this.worldMouseY = mouseY + (camera.y - (camera.height / 2));
+    this.worldMouseX = mouseX + ((int)camera.x - (camera.width / 2));
+    this.worldMouseY = mouseY + ((int)camera.y - (camera.height / 2));
   }
   
 }
